separate java and groovy code to src/main/java and src/main/groovy accordingly
