+++ ch06/AccountManagement/src/main/java/mjg/spring/dao/AccountDAO.java
@@ -1,0 +1,17 @@
+package mjg.spring.dao;
+
+import mjg.spring.entities.Account;
+
+import java.util.List;
+
+public interface AccountDAO {
+    int createAccount(double initialBalance);
+
+    Account findAccountById(int id);
+
+    List<Account> findAllAccounts();
+
+    void updateAccount(Account account);
+
+    void deleteAccount(int id);
+}
