+++ ch06/AccountManagement/src/main/java/mjg/spring/services/AccountProcessor.java
@@ -1,0 +1,27 @@
+package mjg.spring.services;
+
+import mjg.spring.entities.Account;
+
+import java.util.List;
+
+public class AccountProcessor {
+    private List<Account> accounts;
+
+    public void setAccounts(List<Account> accounts) {
+        this.accounts = accounts;
+    }
+
+    public List<Account> getAccounts() {
+        return accounts;
+    }
+
+    public double processAccounts() {
+        double total = 0.0;
+        for (Account account : accounts) {
+            account.withdraw(1.0);
+            total += 1.0;
+        }
+        return total;
+    }
+
+}
